apps/api/requirements.txt
# ============================================
# SENTINANCE API - Python Dependencies
# ============================================
# Install with: pip install -r requirements.txt

# -----------------
# Web Framework
# -----------------
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
qdrant-client>=1.7.0

# -----------------
# Database
# -----------------
sqlalchemy==2.0.23
asyncpg==0.29.0          # Async PostgreSQL driver
alembic==1.13.0          # Database migrations

# -----------------
# Redis
# -----------------
redis==5.0.1

# -----------------
# Authentication
# -----------------
python-jose[cryptography]==3.3.0  # JWT
passlib[bcrypt]==1.7.4            # Password hashing

# -----------------
# Kafka
# -----------------
aiokafka==0.10.0

# -----------------
# AI/ML
# -----------------
google-generativeai>=0.8.0  # Gemini AI
groq>=0.4.0                 # Groq AI (Fallback)
langgraph
numpy>=1.26.0
torch>=2.0.0                # LSTM training
scikit-learn>=1.3.0         # Preprocessing, metrics
yfinance>=0.2.30            # Historical data download

# -----------------
# Data Processing
# -----------------
httpx==0.25.2
beautifulsoup4>=4.12.0    # News scraping
feedparser>=6.0.0         # RSS parsing

# -----------------
# Logging & Observability
# -----------------
structlog==23.2.0
prometheus-client>=0.19.0

# -----------------
# Environment
# -----------------
python-dotenv==1.0.0

# -----------------
# Development
# -----------------
pytest==7.4.3
pytest-asyncio==0.21.1
respx==0.20.2            # HTTP mocking for tests
ruff==0.1.8              # Fast linter
mypy==1.7.1              # Type checking
